Create Sample
In this part of my code, I did not understand how to allow the Arduino to record the time the signal left the sensor, hit the object, and came back to the sensor.  This allowed the correct microseconds to be recorded.  I also needed to look up how to convert from microseconds to centimeters in order to get the correct distance from the sensor to the object being detected.